First pass at migrating heating control to the user defined control scheme. Runs with default settings, no further checks performed
