+++ configs/default.cfg
@@ -216,6 +216,109 @@ e_axis_active = True
 [Temperature Control]
 
 # put temperature control units in here, they won't be accepted anywhere else!
+# Set 'active' = True for all units that are to be parsed. A False value means 
+# it will be ignored
+
+# default PID control for heaters
+# When this config is parsed the 'active' flag is updated to reflect the
+# available thermistors for your Replicape version
+[[Control-E]]
+type = pid-control
+input = Thermistor-E
+pid_Kp = 0.1
+pid_Ti = 100.0
+pid_Td = 0.3
+ok_range = 4.0
+sleep = 0.25
+active = False
+
+[[Control-H]]
+type = pid-control
+input = Thermistor-H
+pid_Kp = 0.1
+pid_Ti = 100.0
+pid_Td = 0.3
+ok_range = 4.0
+sleep = 0.25
+active = False
+
+[[Control-A]]
+type = pid-control
+input = Thermistor-A
+pid_Kp = 0.1
+pid_Ti = 100.0
+pid_Td = 0.3
+ok_range = 4.0
+sleep = 0.25
+active = False
+
+[[Control-B]]
+type = pid-control
+input = Thermistor-B
+pid_Kp = 0.1
+pid_Ti = 100.0
+pid_Td = 0.3
+ok_range = 4.0
+sleep = 0.25
+active = False
+
+[[Control-C]]
+type = pid-control
+input = Thermistor-C
+pid_Kp = 0.1
+pid_Ti = 100.0
+pid_Td = 0.3
+ok_range = 4.0
+sleep = 0.25
+active = False
+
+[[Control-HBP]]
+type = pid-control
+input = Thermistor-HBP
+pid_Kp = 0.1
+pid_Ti = 100.0
+pid_Td = 0.3
+ok_range = 4.0
+sleep = 0.5
+active = False
+
+[Thermistors]
+
+# Thermistors for measuring temperature
+# For list of available temp charts, look in temp_chart.py
+# When this config is parsed the 'active' flag is updated to reflect the
+# available thermistors for your Replicape version
+
+[[Thermistor-E]]
+sensor = B57560G104F
+path_adc = /sys/bus/iio/devices/iio:device0/in_voltage4_raw
+active = False
+
+[[Thermistor-H]]
+sensor = B57560G104F
+path_adc = /sys/bus/iio/devices/iio:device0/in_voltage5_raw
+active = False
+
+[[Thermistor-A]]
+sensor = B57560G104F
+path_adc = /sys/bus/iio/devices/iio:device0/in_voltage0_raw
+active = False
+
+[[Thermistor-B]]
+sensor = B57560G104F
+path_adc = /sys/bus/iio/devices/iio:device0/in_voltage3_raw
+active = False
+
+[[Thermistor-C]]
+sensor = B57560G104F
+path_adc = /sys/bus/iio/devices/iio:device0/in_voltage2_raw
+active = False
+
+[[Thermistor-HBP]]
+sensor = B57560G104F
+path_adc = /sys/bus/iio/devices/iio:device0/in_voltage6_raw
+active = False
+
 
 [Fans]
 
@@ -228,118 +331,127 @@ type = fan
 input = 0
 add-to-M106 = False
 channel = 0
+active = True
 
 [[Fan-1]]
 type=fan
 input = 0
 add-to-M106 = False
 channel = 0
+active = True
 
 [[Fan-2]]
 type = fan
 input = 0
 add-to-M106 = False
 channel = 0
+active = True
 
 [[Fan-3]]
 type = fan
 input = 0
 add-to-M106 = False
 channel = 0
+active = True
 
 
 [Heaters]
-# For list of available temp charts, look in temp_chart.py
 
-sensor_E = B57560G104F
-pid_Kp_E = 0.1
-pid_Ti_E = 100.0
-pid_Td_E = 0.3
-ok_range_E = 4.0
-max_rise_temp_E = 10.0
-max_fall_temp_E = 10.0
-min_temp_E = 20.0
-max_temp_E = 250.0
-path_adc_E = /sys/bus/iio/devices/iio:device0/in_voltage4_raw
-mosfet_E = 5
-onoff_E = False
-prefix_E = T0
-max_power_E = 1.0
+# make things hot, can be connected to control units defined in [Temperature Control]
+# NOTE: Safety parameters are important, make sure they are set correctly for your printer!
+# When this config is parsed the 'active' flag is updated to reflect the
+# available heaters for your Replicape version
 
-sensor_H = B57560G104F
-pid_Kp_H = 0.1
-pid_Ti_H = 0.01
-pid_Td_H = 0.3
-ok_range_H = 4.0
-max_rise_temp_H = 10.0
-max_fall_temp_H = 10.0
-min_temp_H = 20.0
-max_temp_H = 250.0
-path_adc_H = /sys/bus/iio/devices/iio:device0/in_voltage5_raw
-mosfet_H = 3
-onoff_H = False
-prefix_H = T1
-max_power_H = 1.0
+[[Heater-E]]
+type = heater
+mosfet = 5
+max_rise_temp = 10.0
+min_rise_temp = 0.1
+max_fall_temp = 10.0
+min_temp = 20.0
+max_temp = 250.0
+max_power = 1.0
+sleep = 0.25
+prefix = T0
+temperature = Thermistor-E
+#input = 0
+input = Control-E
+active = False
 
-sensor_A = B57560G104F
-pid_Kp_A = 0.1
-pid_Ti_A = 0.01
-pid_Td_A = 0.3
-ok_range_A = 4.0
-max_rise_temp_A = 10.0
-max_fall_temp_A = 10.0
-min_temp_A = 20.0
-max_temp_A = 250.0
-path_adc_A = /sys/bus/iio/devices/iio:device0/in_voltage0_raw
-mosfet_A = 11
-onoff_A = False
-prefix_A = T2
-max_power_A = 1.0
+[[Heater-H]]
+type = heater
+mosfet = 3
+max_rise_temp = 10.0
+min_rise_temp = 0.1
+max_fall_temp = 10.0
+min_temp = 20.0
+max_temp = 250.0
+max_power = 1.0
+sleep = 0.25
+prefix = T1
+temperature = Thermistor-H
+input = Control-H
+active = False
 
-sensor_B = B57560G104F
-pid_Kp_B = 0.1
-pid_Ti_B = 0.01
-pid_Td_B = 0.3
-ok_range_B = 4.0
-max_rise_temp_B = 10.0
-max_fall_temp_B = 10.0
-min_temp_B = 20.0
-max_temp_B = 250.0
-path_adc_B = /sys/bus/iio/devices/iio:device0/in_voltage3_raw
-mosfet_B = 12
-onoff_B = False
-prefix_B = T3
-max_power_B = 1.0
+[[Heater-A]]
+type = heater
+mosfet = 11
+max_rise_temp = 10.0
+min_rise_temp = 0.1
+max_fall_temp = 10.0
+min_temp = 20.0
+max_temp = 250.0
+max_power = 1.0
+sleep = 0.25
+prefix = T2
+temperature = Thermistor-A
+input = Control-A
+active = False
 
-sensor_C = B57560G104F
-pid_Kp_C = 0.1
-pid_Ti_C = 0.01
-pid_Td_C = 0.3
-ok_range_C = 4.0
-max_rise_temp_C = 10.0
-max_fall_temp_C = 10.0
-min_temp_C = 20.0
-max_temp_C = 250.0
-path_adc_C = /sys/bus/iio/devices/iio:device0/in_voltage2_raw
-mosfet_C = 13
-onoff_C = False
-prefix_C = T4
-max_power_C = 1.0
+[[Heater-B]]
+type = heater
+mosfet = 12
+max_rise_temp = 10.0
+min_rise_temp = 0.1
+max_fall_temp = 10.0
+min_temp = 20.0
+max_temp = 250.0
+max_power = 1.0
+sleep = 0.25
+prefix = T3
+temperature = Thermistor-B
+input = Control-B
+active = False
 
-sensor_HBP = B57560G104F
-pid_Kp_HBP = 0.1
-pid_Ti_HBP = 0.01
-pid_Td_HBP = 0.3
-ok_range_HBP = 4.0
-max_rise_temp_HBP = 10.0
-max_fall_temp_HBP = 10.0
-min_temp_HBP = 20.0
-max_temp_HBP = 250.0
-path_adc_HBP = /sys/bus/iio/devices/iio:device0/in_voltage6_raw
-mosfet_HBP = 4
-onoff_HBP = False
-prefix_HBP = B
-max_power_HBP = 1.0
+[[Heater-C]]
+type = heater
+mosfet = 13
+max_rise_temp = 10.0
+min_rise_temp = 0.1
+max_fall_temp = 10.0
+min_temp = 20.0
+max_temp = 250.0
+max_power = 1.0
+sleep = 0.25
+prefix = T4
+temperature = Thermistor-C
+input = Control-C
+active = False
+
+[[Heater-HBP]]
+type = heater
+mosfet = 4
+max_rise_temp = 10.0
+min_rise_temp = 0.1
+max_fall_temp = 10.0
+min_temp = 20.0
+max_temp = 250.0
+max_power = 1.0
+sleep = 0.5
+prefix = B
+temperature = Thermistor-HBP
+input = Control-HBP
+active = False
 
 [Endstops]
 # Which axis should be homed.
